Download data package

Download all data related to an animal project as a data package that can be deposited in a research data repository. Includes option to filter on scientific names.

download_dataset(
  connection = con,
  animal_project_code,
  scientific_name = NULL,
  directory = animal_project_code
)

Arguments

connection
A connection to the ETN database. Defaults to con.

animal_project_code
Character. Animal project you want to download data for. Required.

scientific_name
Character (vector). One or more scientific names. Defaults to no filter (i.e. include all scientific names, including reference tags).

directory
Character. Relative path to local download directory. Defaults to creating a directory named after animal project code. Existing files of the same name will be overwritten.

Details

The data are downloaded as a Frictionless Data Package containing:

file	description

animals.csv	Animals related to an animal_project_code, optionally filtered on scientific_name(s), as returned by get_animals().

tags.csv	Tags associated with the selected animals, as returned by get_tags().

detections.csv	Detections for the selected animals, as returned by get_detections().

deployments.csv	Deployments for the network_project_code(s) found in detections, as returned by get_deployments(). This allows users to see when receivers were deployed, even if these did not detect the selected animals.

receivers.csv	Receivers for the selected deployments, as returned by get_receivers().

datapackage.json	A Frictionless Table Schema metadata file describing the fields and relations of the above csv files. This file is copied from here and can be used to validate the data package.

The function will report the number of records per csv file, as well as the included scientific names and network projects. Warnings will be raised for:

Animals with multiple tags

Tags associated with multiple animals

Deployments without network project: these deployments will not be listed in deployments.csv and will therefore raise a foreign key validation error.

Duplicate detections: detections with the duplicate pk. These are removed by the function in detections.csv.

Important: The data are downloaded as is from the database, i.e. no quality or consistency checks are performed by this function. We therefore recommend to verify the data before publication. A consistency check can be performed by validation tools of the Frictionless Framework, e.g. frictionless validate datapackage.json on the command line using frictionless-py.

Examples

if (FALSE) {
# Download data for the 2014_demer animal project (all scientific names)
download_dataset(con, animal_project_code = "2014_demer")
#> Downloading data to directory "2014_demer":
#> * (1/6): downloading animals.csv
#> * (2/6): downloading tags.csv
#> * (3/6): downloading detections.csv
#> * (4/6): downloading deployments.csv
#> * (5/6): downloading receivers.csv
#> * (6/6): adding datapackage.json as file metadata
#>
#> Summary statistics for dataset "2014_demer":
#> * number of animals:           16
#> * number of tags:              16
#> * number of detections:        237064
#> * number of deployments:       939
#> * number of receivers:         179
#> * included scientific names:   Petromyzon marinus, Rutilus rutilus, Silurus glanis, Squalius cephalus
#> * included network projects:   albert, demer, dijle, no_info, zeeschelde
}