Get animal project data

Get data for animal projects, with options to filter results.

Usage

get_animal_projects(connection, animal_project_code = NULL, citation = FALSE)

Arguments

connection

A connection to the ETN database. This argument is no longer used. You will be prompted for credentials instead.

animal_project_code

Character (vector). One or more animal project codes. Case-insensitive.

citation

Logical. If TRUE, adds extra columns with citation information for each project from MarineInfo using the imis_dataset_id:

citation: Formatted citation with DOI if available.

doi: DOI for the dataset if available.

contact_name: Contact person, usually the first author. If no contact person is provided, the first author with status creator.

contact_email: Email address of the contact person.

contact_affiliation: Institute of the contact person.

Value

A tibble with animal project data, sorted by project_code.

See also

Examples

# Get all animal projects
get_animal_projects()
#> Error in httr2::req_perform(httr2::req_body_json(httr2::req_url_path_append(httr2::request(domain),     "validate_login", "json/"), data = credentials)): HTTP 503 Service Unavailable.

# Get a specific animal project with citation
get_animal_projects(animal_project_code = "2014_demer", citation = TRUE)
#> Error in httr2::req_perform(httr2::req_body_json(httr2::req_url_path_append(httr2::request(domain),     "validate_login", "json/"), data = credentials)): HTTP 503 Service Unavailable.