Get animal project data as a Data Package

Gets all data related to an animal project as a Data Package.

Usage

get_package(animal_project_code)

Arguments

animal_project_code

Animal project you want to get data from.

Value

A Data Package object. Write it to disk with write_package().

Included resources

The Data Package will contain and describe the following resources:

animals: Animals related to an animal_project_code, as returned by get_animals().

tags: Tags associated with the selected animals, as returned by get_tags().

detections: Acoustic detections for the selected animals, as returned by get_acoustic_detections().

deployments: Acoustic deployments for the acoustic_project_code(s) found in detections, as returned by get_acoustic_deployments(). This allows you to see when receivers were deployed, even if these did not detect the selected animals.

receivers: Acoustic receivers for the selected deployments, as returned by get_acoustic_receivers().

Data quality

The data are downloaded from the ETN database as is, i.e. no quality or consistency checks are performed. Verifying the data before publication is therefore recommended. You can validate the technical consistency of your Data Package using Frictionless Framework with:

pip install frictionless
frictionless validate datapackage.json

See also

Examples

if (FALSE) { # etn:::credentials_are_set() & interactive()
# Get a Data Package for a project
get_package(animal_project_code = "2014_demer")
}