Support function to get unique set of scientific names

This function retrieves all unique scientific names

scientific_names(connection)

Arguments

connection
A valid connection to ETN database.

Value

A vector of all scientific names present in vliz.animals_view.