Get acoustic project data

Get data for acoustic projects, with options to filter results.

Usage

get_acoustic_projects(connection = con, acoustic_project_code = NULL)

Arguments

connection

A connection to the ETN database. Defaults to con.

acoustic_project_code

Character (vector). One or more acoustic project codes. Case-insensitive.

Value

A tibble with acoustic project data, sorted by project_code. See also field definitions.

Examples

# Set default connection variable
con <- connect_to_etn()

# Get all acoustic projects
get_acoustic_projects(con)
#> # A tibble: 244 × 11
#>    project_id project_code   project_type telemetry_type project_name start_date
#>         <int> <chr>          <chr>        <chr>          <chr>        <date>    
#>  1        794 2004_Gudena    acoustic     Acoustic       2004_Gudena  2004-01-01
#>  2        761 2011_Loire     acoustic     Acoustic       2011_Loire   2011-11-11
#>  3        755 2011_Warnow    acoustic     Acoustic       2011_Warnow  2011-06-01
#>  4         11 2011_bovensch… acoustic     Acoustic       2011_Bovens… 2011-01-01
#>  5        758 2013_Foyle     acoustic     Acoustic       2013_Foyle   2013-07-01
#>  6         13 2013_Maas      acoustic     Acoustic       2013_Maas    2013-01-01
#>  7        802 2014_Frome     acoustic     Acoustic       2014_Frome   2014-10-01
#>  8        819 2014_Nene      acoustic     Acoustic       2014_Nene    2014-10-31
#>  9        815 2015_PhD_Gutm… acoustic     Acoustic       2015_PhD_Gu… 2015-09-22
#> 10        773 2016_Diaccia_… acoustic     PIT            2016_Diacci… 2016-01-01
#> # ℹ 234 more rows
#> # ℹ 5 more variables: end_date <date>, latitude <dbl>, longitude <dbl>,
#> #   moratorium <lgl>, imis_dataset_id <int>

# Get a specific acoustic project
get_acoustic_projects(con, acoustic_project_code = "demer")
#> # A tibble: 1 × 11
#>   project_id project_code project_type telemetry_type project_name start_date
#>        <int> <chr>        <chr>        <chr>          <chr>        <date>    
#> 1          7 demer        acoustic     Acoustic       Demer        2014-04-10
#> # ℹ 5 more variables: end_date <date>, latitude <dbl>, longitude <dbl>,
#> #   moratorium <lgl>, imis_dataset_id <int>