Get acoustic project data

Get data for acoustic projects, with options to filter results.

Usage

get_acoustic_projects(connection = con, acoustic_project_code = NULL)

Arguments

connection

A connection to the ETN database. Defaults to con.

acoustic_project_code

Character (vector). One or more acoustic project codes. Case-insensitive.

Value

A tibble with acoustic project data, sorted by project_code. See also field definitions.

Examples

# Set default connection variable
con <- connect_to_etn()

# Get all acoustic projects
get_acoustic_projects(con)
#> # A tibble: 194 × 11
#>    proje…¹ proje…² proje…³ telem…⁴ proje…⁵ start_date end_date   latit…⁶ longi…⁷
#>      <int> <chr>   <chr>   <chr>   <chr>   <date>     <date>       <dbl>   <dbl>
#>  1     794 2004_G… acoust… Acoust… 2004_G… 2004-01-01 2005-12-31    56.4   9.91 
#>  2     761 2011_L… acoust… Acoust… 2011_L… 2011-11-11 2012-02-25    47.3  -1.98 
#>  3     755 2011_W… acoust… Acoust… 2011_W… 2011-06-01 2012-10-12    54.1  12.1  
#>  4     758 2013_F… acoust… Acoust… 2013_F… 2013-07-01 2014-03-01    54.9  -7.39 
#>  5      13 2013_M… acoust… Acoust… 2013_M… 2013-01-01 2017-12-31    50.7   5.62 
#>  6     802 2014_F… acoust… Acoust… 2014_F… 2014-10-01 2015-01-31    50.7  -2.11 
#>  7     819 2014_N… acoust… Acoust… 2014_N… 2014-10-31 2015-02-01    52.5  -0.457
#>  8     815 2015_P… acoust… Acoust… 2015_P… 2015-09-22 2017-09-07    52.2  -2.24 
#>  9     773 2016_D… acoust… PIT     2016_D… 2016-01-01 NA            42.8  10.9  
#> 10     763 2017_F… acoust… Acoust… 2017_F… 2017-09-10 2018-02-23    48.6  -2.08 
#> # … with 184 more rows, 2 more variables: moratorium <chr>,
#> #   imis_dataset_id <int>, and abbreviated variable names ¹​project_id,
#> #   ²​project_code, ³​project_type, ⁴​telemetry_type, ⁵​project_name, ⁶​latitude,
#> #   ⁷​longitude

# Get a specific acoustic project
get_acoustic_projects(con, acoustic_project_code = "demer")
#> # A tibble: 1 × 11
#>   projec…¹ proje…² proje…³ telem…⁴ proje…⁵ start_date end_date   latit…⁶ longi…⁷
#>      <int> <chr>   <chr>   <chr>   <chr>   <date>     <date>       <dbl>   <dbl>
#> 1        7 demer   acoust… Acoust… Demer   2014-04-10 2015-02-13      NA      NA
#> # … with 2 more variables: moratorium <chr>, imis_dataset_id <int>, and
#> #   abbreviated variable names ¹​project_id, ²​project_code, ³​project_type,
#> #   ⁴​telemetry_type, ⁵​project_name, ⁶​latitude, ⁷​longitude