Get acoustic receiver data

Get data for acoustic receivers, with options to filter results.

Usage

get_acoustic_receivers(connection = con, receiver_id = NULL, status = NULL)

Arguments

connection

A connection to the ETN database. Defaults to con.

receiver_id

Character (vector). One or more receiver identifiers.

status

Character. One or more statuses, e.g. available or broken.

Value

A tibble with acoustic receiver data, sorted by receiver_id. See also field definitions. Values for owner_organization will only be visible if you are member of the group.

Examples

# Set default connection variable
con <- connect_to_etn()

# Get all acoustic receivers
get_acoustic_receivers(con)
#> # A tibble: 4,849 × 23
#>    receiver_id    manuf…¹ recei…² recei…³ modem…⁴ status batte…⁵ owner…⁶ finan…⁷
#>    <chr>          <chr>   <chr>   <chr>   <chr>   <chr>  <chr>   <chr>   <chr>  
#>  1 ADTT-LP-7-3-3  THELMA… ADTT    LP-7-3… NA      avail… NA      OTN     V2LIMF…
#>  2 ADTT-LP-7-3-4  THELMA… ADTT    LP-7-3… NA      avail… NA      OTN     V2LIMF…
#>  3 BULK-UPLOAD-D… VEMCO   BULK    UPLOAD… NA      avail… NA      VLIZ    NA     
#>  4 CURRENT METER… AANDER… CURREN… Aander… NA      NA     NA      OTN     SGB    
#>  5 DST TILT-111   STAR-O… DST TI… 111     NA      NA     NA      OTN     SGB    
#>  6 DST TILT-112   STAR-O… DST TI… 112     NA      NA     NA      OTN     SGB    
#>  7 DST TILT-113   STAR-O… DST TI… 113     NA      NA     NA      OTN     SGB    
#>  8 DST TILT-202   STAR-O… DST TI… 202     NA      NA     NA      OTN     SGB    
#>  9 DST TILT-205   STAR-O… DST TI… 205     NA      NA     NA      OTN     SGB    
#> 10 DST TILT-207   STAR-O… DST TI… 207     NA      NA     NA      OTN     SGB    
#> # … with 4,839 more rows, 14 more variables: built_in_acoustic_tag_id <chr>,
#> #   ar_model <chr>, ar_serial_number <chr>, ar_battery_estimated_life <chr>,
#> #   ar_voltage_at_deploy <chr>, ar_interrogate_code <chr>,
#> #   ar_receive_frequency <chr>, ar_reply_frequency <chr>, ar_ping_rate <chr>,
#> #   ar_enable_code_address <chr>, ar_release_code <chr>, ar_disable_code <chr>,
#> #   ar_tilt_code <chr>, ar_tilt_after_deploy <chr>, and abbreviated variable
#> #   names ¹​manufacturer, ²​receiver_model, ³​receiver_serial_number, …

# Get lost and broken acoustic receivers
get_acoustic_receivers(con, status = c("lost", "broken"))
#> # A tibble: 161 × 23
#>    receiver_id    manuf…¹ recei…² recei…³ modem…⁴ status batte…⁵ owner…⁶ finan…⁷
#>    <chr>          <chr>   <chr>   <chr>   <chr>   <chr>  <chr>   <chr>   <chr>  
#>  1 TBR700-001195  THELMA… TBR700  001195  NA      lost   8       PROFISH Wallon…
#>  2 TBR700-001197  THELMA… TBR700  001197  NA      lost   8       PROFISH Wallon…
#>  3 TBR700-001768  THELMA… TBR700  001768  NA      broken 8       PROFISH Wallon…
#>  4 TBR700-TBR700… THELMA… TBR700  TBR700… NA      broken 8       PROFISH Wallon…
#>  5 VR2AR-545719   VEMCO   VR2AR   545719  NA      lost   NA      VLIZ    NA     
#>  6 VR2AR-546046   VEMCO   VR2AR   546046  NA      lost   NA      VLIZ    no_info
#>  7 VR2AR-546049   VEMCO   VR2AR   546049  NA      lost   NA      VLIZ    NA     
#>  8 VR2AR-546050   VEMCO   VR2AR   546050  NA      lost   NA      VLIZ    NA     
#>  9 VR2AR-546894   VEMCO   VR2AR   546894  NA      lost   NA      VLIZ    NA     
#> 10 VR2AR-547656   VEMCO   VR2AR   547656  NA      lost   NA      VLIZ    NA     
#> # … with 151 more rows, 14 more variables: built_in_acoustic_tag_id <chr>,
#> #   ar_model <chr>, ar_serial_number <chr>, ar_battery_estimated_life <chr>,
#> #   ar_voltage_at_deploy <chr>, ar_interrogate_code <chr>,
#> #   ar_receive_frequency <chr>, ar_reply_frequency <chr>, ar_ping_rate <chr>,
#> #   ar_enable_code_address <chr>, ar_release_code <chr>, ar_disable_code <chr>,
#> #   ar_tilt_code <chr>, ar_tilt_after_deploy <chr>, and abbreviated variable
#> #   names ¹​manufacturer, ²​receiver_model, ³​receiver_serial_number, …

# Get a specific acoustic receiver
get_acoustic_receivers(con, receiver_id = "VR2W-124070")
#> # A tibble: 1 × 23
#>   recei…¹ manuf…² recei…³ recei…⁴ modem…⁵ status batte…⁶ owner…⁷ finan…⁸ built…⁹
#>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  <chr>   <chr>   <chr>   <chr>  
#> 1 VR2W-1… VEMCO   VR2W    124070  NA      avail… NA      INBO    NA      NA     
#> # … with 13 more variables: ar_model <chr>, ar_serial_number <chr>,
#> #   ar_battery_estimated_life <chr>, ar_voltage_at_deploy <chr>,
#> #   ar_interrogate_code <chr>, ar_receive_frequency <chr>,
#> #   ar_reply_frequency <chr>, ar_ping_rate <chr>, ar_enable_code_address <chr>,
#> #   ar_release_code <chr>, ar_disable_code <chr>, ar_tilt_code <chr>,
#> #   ar_tilt_after_deploy <chr>, and abbreviated variable names ¹​receiver_id,
#> #   ²​manufacturer, ³​receiver_model, ⁴​receiver_serial_number, ⁵​modem_address, …