Get animal project data

Get data for animal projects, with options to filter results.

get_animal_projects(connection = con, animal_project_code = NULL)

Arguments

connection
A connection to the ETN database. Defaults to con.

animal_project_code
Character (vector). One or more animal project codes. Case-insensitive.

Value

A tibble with animal project data, sorted by project_code. See also field definitions.

Examples

# Set default connection variable
con <- connect_to_etn()

# Get all animal projects
get_animal_projects(con)

#> # A tibble: 184 × 11
#>    project_id project_code  project_type telemetry_type project_name  start_date
#>         <int> <chr>         <chr>        <chr>          <chr>         <date>    
#>  1        793 2004_Gudena   animal       Acoustic       2004_Gudena   2004-01-01
#>  2         16 2010_phd_reu… animal       Acoustic       2010_phd_reu… 2010-08-01
#>  3        841 2010_phd_reu… animal       Acoustic       2010_phd_reu… 2010-08-01
#>  4        760 2011_Loire    animal       Acoustic       2011_Loire    2011-11-11
#>  5         20 2011_rivierp… animal       Acoustic       2011 Rivierp… 2011-01-01
#>  6        754 2011_Warnow   animal       Acoustic       2011_Warnow   2011-06-01
#>  7         15 2012_leopold… animal       Acoustic       2012 Leopold… 2012-01-01
#>  8         18 2013_albertk… animal       Acoustic       2013 Albertk… 2013-10-10
#>  9        757 2013_Foyle    animal       Acoustic       2013_Foyle    2013-07-01
#> 10         21 2014_demer    animal       Acoustic       2014 Demer    2014-04-10
#> # … with 174 more rows, and 5 more variables: end_date <date>, latitude <dbl>,
#> #   longitude <dbl>, moratorium <chr>, imis_dataset_id <int>

# Get a specific animal project
get_animal_projects(con, animal_project_code = "2014_demer")

#> # A tibble: 1 × 11
#>   project_id project_code project_type telemetry_type project_name start_date
#>        <int> <chr>        <chr>        <chr>          <chr>        <date>    
#> 1         21 2014_demer   animal       Acoustic       2014 Demer   2014-04-10
#> # … with 5 more variables: end_date <date>, latitude <dbl>, longitude <dbl>,
#> #   moratorium <chr>, imis_dataset_id <int>