Get animal project data

Get data for animal projects, with options to filter results.

Usage

get_animal_projects(connection = con, animal_project_code = NULL)

Arguments

connection

A connection to the ETN database. Defaults to con.

animal_project_code

Character (vector). One or more animal project codes. Case-insensitive.

Value

A tibble with animal project data, sorted by project_code. See also field definitions.

Examples

# Set default connection variable
con <- connect_to_etn()

# Get all animal projects
get_animal_projects(con)
#> # A tibble: 218 × 11
#>    proje…¹ proje…² proje…³ telem…⁴ proje…⁵ start_date end_date   latit…⁶ longi…⁷
#>      <int> <chr>   <chr>   <chr>   <chr>   <date>     <date>       <dbl>   <dbl>
#>  1     793 2004_G… animal  Acoust… 2004_G… 2004-01-01 2005-12-31    56.4    9.91
#>  2      16 2010_p… animal  Acoust… 2010_p… 2010-08-01 2012-10-30    NA     NA   
#>  3     841 2010_p… animal  Acoust… 2010_p… 2010-08-01 2012-10-30    NA     NA   
#>  4     760 2011_L… animal  Acoust… 2011_L… 2011-11-11 2012-02-25    47.3   -1.98
#>  5     754 2011_W… animal  Acoust… 2011_W… 2011-06-01 2012-10-12    54.1   12.1 
#>  6      20 2011_r… animal  Acoust… 2011 R… 2011-01-01 2012-09-03    NA     NA   
#>  7      15 2012_l… animal  Acoust… 2012 L… 2012-01-01 2016-01-18    NA     NA   
#>  8     757 2013_F… animal  Acoust… 2013_F… 2013-07-01 2014-03-01    54.9   -7.39
#>  9      18 2013_a… animal  Acoust… 2013 A… 2013-10-10 2018-12-31    NA     NA   
#> 10     801 2014_F… animal  Acoust… 2014_F… 2014-10-01 2015-01-31    50.7   -2.11
#> # … with 208 more rows, 2 more variables: moratorium <chr>,
#> #   imis_dataset_id <int>, and abbreviated variable names ¹​project_id,
#> #   ²​project_code, ³​project_type, ⁴​telemetry_type, ⁵​project_name, ⁶​latitude,
#> #   ⁷​longitude

# Get a specific animal project
get_animal_projects(con, animal_project_code = "2014_demer")
#> # A tibble: 1 × 11
#>   projec…¹ proje…² proje…³ telem…⁴ proje…⁵ start_date end_date   latit…⁶ longi…⁷
#>      <int> <chr>   <chr>   <chr>   <chr>   <date>     <date>       <dbl>   <dbl>
#> 1       21 2014_d… animal  Acoust… 2014 D… 2014-04-10 2015-02-13      NA      NA
#> # … with 2 more variables: moratorium <chr>, imis_dataset_id <int>, and
#> #   abbreviated variable names ¹​project_id, ²​project_code, ³​project_type,
#> #   ⁴​telemetry_type, ⁵​project_name, ⁶​latitude, ⁷​longitude