Get tag data

Get data for tags, with options to filter results. Note that there can be multiple records (acoustic_tag_id) per tag device (tag_serial_number).

Usage

get_tags(
  connection = con,
  tag_type = NULL,
  tag_subtype = NULL,
  tag_serial_number = NULL,
  acoustic_tag_id = NULL
)

Arguments

connection

A connection to the ETN database. Defaults to con.

tag_type

Character (vector). acoustic or archival. Some tags are both, find those with acoustic-archival.

tag_subtype

Character (vector). animal, built-in, range or sentinel.

tag_serial_number

Character (vector). One or more tag serial numbers.

acoustic_tag_id

Character (vector). One or more acoustic tag identifiers, i.e. identifiers found in get_acoustic_detections().

Value

A tibble with tags data, sorted by tag_serial_number. See also field definitions. Values for owner_organization and owner_pi will only be visible if you are member of the group.

Examples

# Set default connection variable
con <- connect_to_etn()

# Get all tags
get_tags(con)
#> # A tibble: 22,882 × 54
#>    tag_se…¹ tag_t…² tag_s…³ senso…⁴ acous…⁵ acous…⁶ manuf…⁷ model frequ…⁸ status
#>    <chr>    <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr> <chr>   <chr> 
#>  1 10       acoust… animal  NA      S256-10 A69-11… THELMA… LP6-D 69 kHz  active
#>  2 11       acoust… animal  NA      S256-11 A69-11… THELMA… LP6-D 69 kHz  active
#>  3 12       acoust… animal  NA      S256-12 A69-11… THELMA… LP6-D 69 kHz  active
#>  4 13       acoust… animal  NA      S256-13 A69-11… THELMA… LP6-D 69 kHz  active
#>  5 14       acoust… animal  NA      S256-14 A69-11… THELMA… LP6-D 69 kHz  active
#>  6 15       acoust… animal  NA      S256-15 A69-11… THELMA… LP6-D 69 kHz  active
#>  7 16       acoust… animal  NA      S256-16 A69-11… THELMA… LP6-D 69 kHz  active
#>  8 17       acoust… animal  NA      S256-17 A69-11… THELMA… LP6-D 69 kHz  active
#>  9 18       acoust… animal  NA      S256-18 A69-11… THELMA… LP6-D 69 kHz  active
#> 10 19       acoust… animal  NA      S256-19 A69-11… THELMA… LP6-D 69 kHz  active
#> # … with 22,872 more rows, 44 more variables: activation_date <dttm>,
#> #   battery_estimated_life <chr>, battery_estimated_end_date <dttm>,
#> #   length <dbl>, diameter <dbl>, weight <dbl>, floating <chr>,
#> #   archive_memory <chr>, sensor_slope <dbl>, sensor_intercept <dbl>,
#> #   sensor_range <chr>, sensor_range_min <dbl>, sensor_range_max <dbl>,
#> #   sensor_resolution <dbl>, sensor_unit <chr>, sensor_accuracy <dbl>,
#> #   sensor_transmit_ratio <int>, accelerometer_algorithm <chr>, …

# Get archival tags, including acoustic-archival
get_tags(con, tag_type = c("archival", "acoustic-archival"))
#> # A tibble: 3,826 × 54
#>    tag_se…¹ tag_t…² tag_s…³ senso…⁴ acous…⁵ acous…⁶ manuf…⁷ model frequ…⁸ status
#>    <chr>    <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr> <chr>   <chr> 
#>  1 4879H    acoust… animal  NA      A69-13… R64K-5… VEMCO   V13-… NA      avail…
#>  2 4880H    acoust… animal  NA      A69-13… R64K-5… VEMCO   V13-… NA      avail…
#>  3 4882H    acoust… animal  NA      A69-13… R64K-5… VEMCO   V13-… NA      avail…
#>  4 31537    acoust… animal  NA      A69-16… NA      VEMCO   V13-… NA      active
#>  5 31538    acoust… animal  NA      A69-16… NA      VEMCO   V13-… NA      active
#>  6 31539    acoust… animal  NA      A69-16… NA      VEMCO   V13-… NA      active
#>  7 31540    acoust… animal  NA      A69-16… NA      VEMCO   V13-… NA      active
#>  8 31541    acoust… animal  NA      A69-16… NA      VEMCO   V13-… NA      active
#>  9 31542    acoust… animal  NA      A69-16… NA      VEMCO   V13-… NA      active
#> 10 31543    acoust… animal  NA      A69-16… NA      VEMCO   V13-… NA      active
#> # … with 3,816 more rows, 44 more variables: activation_date <dttm>,
#> #   battery_estimated_life <chr>, battery_estimated_end_date <dttm>,
#> #   length <dbl>, diameter <dbl>, weight <dbl>, floating <chr>,
#> #   archive_memory <chr>, sensor_slope <dbl>, sensor_intercept <dbl>,
#> #   sensor_range <chr>, sensor_range_min <dbl>, sensor_range_max <dbl>,
#> #   sensor_resolution <dbl>, sensor_unit <chr>, sensor_accuracy <dbl>,
#> #   sensor_transmit_ratio <int>, accelerometer_algorithm <chr>, …

# Get tags of specific subtype
get_tags(con, tag_subtype = c("built-in", "range"))
#> # A tibble: 299 × 54
#>    tag_se…¹ tag_t…² tag_s…³ senso…⁴ acous…⁵ acous…⁶ manuf…⁷ model frequ…⁸ status
#>    <chr>    <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr> <chr>   <chr> 
#>  1 2060     acoust… built-… NA      R01M-8… NA      THELMA… TBR8… 069k    avail…
#>  2 2061     acoust… built-… NA      R01M-8… NA      THELMA… TBR8… 069k    avail…
#>  3 2063     acoust… built-… NA      R01M-8… NA      THELMA… TBR8… 069k    avail…
#>  4 2066     acoust… built-… NA      R01M-8… NA      THELMA… TBR8… 069k    avail…
#>  5 2067     acoust… built-… NA      R01M-8… NA      THELMA… TBR8… 069k    avail…
#>  6 2068     acoust… built-… NA      R01M-8… NA      THELMA… TBR8… 069k    avail…
#>  7 2070     acoust… built-… NA      R01M-8… NA      THELMA… TBR8… 069k    avail…
#>  8 2071     acoust… built-… NA      R01M-8… NA      THELMA… TBR8… 069k    avail…
#>  9 2072     acoust… built-… NA      R01M-8… NA      THELMA… TBR8… 069k    avail…
#> 10 2073     acoust… built-… NA      R01M-8… NA      THELMA… TBR8… 069k    avail…
#> # … with 289 more rows, 44 more variables: activation_date <dttm>,
#> #   battery_estimated_life <chr>, battery_estimated_end_date <dttm>,
#> #   length <dbl>, diameter <dbl>, weight <dbl>, floating <chr>,
#> #   archive_memory <chr>, sensor_slope <dbl>, sensor_intercept <dbl>,
#> #   sensor_range <chr>, sensor_range_min <dbl>, sensor_range_max <dbl>,
#> #   sensor_resolution <dbl>, sensor_unit <chr>, sensor_accuracy <dbl>,
#> #   sensor_transmit_ratio <int>, accelerometer_algorithm <chr>, …

# Get specific tags (note that these can return multiple records)
get_tags(con, tag_serial_number = "1187450")
#> # A tibble: 1 × 54
#>   tag_ser…¹ tag_t…² tag_s…³ senso…⁴ acous…⁵ acous…⁶ manuf…⁷ model frequ…⁸ status
#>   <chr>     <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr> <chr>   <chr> 
#> 1 1187450   acoust… animal  NA      A69-16… NA      VEMCO   V7-4x 069k    avail…
#> # … with 44 more variables: activation_date <dttm>,
#> #   battery_estimated_life <chr>, battery_estimated_end_date <dttm>,
#> #   length <dbl>, diameter <dbl>, weight <dbl>, floating <chr>,
#> #   archive_memory <chr>, sensor_slope <dbl>, sensor_intercept <dbl>,
#> #   sensor_range <chr>, sensor_range_min <dbl>, sensor_range_max <dbl>,
#> #   sensor_resolution <dbl>, sensor_unit <chr>, sensor_accuracy <dbl>,
#> #   sensor_transmit_ratio <int>, accelerometer_algorithm <chr>, …
get_tags(con, acoustic_tag_id = "A69-1601-16130")
#> # A tibble: 1 × 54
#>   tag_ser…¹ tag_t…² tag_s…³ senso…⁴ acous…⁵ acous…⁶ manuf…⁷ model frequ…⁸ status
#>   <chr>     <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr> <chr>   <chr> 
#> 1 1187450   acoust… animal  NA      A69-16… NA      VEMCO   V7-4x 069k    avail…
#> # … with 44 more variables: activation_date <dttm>,
#> #   battery_estimated_life <chr>, battery_estimated_end_date <dttm>,
#> #   length <dbl>, diameter <dbl>, weight <dbl>, floating <chr>,
#> #   archive_memory <chr>, sensor_slope <dbl>, sensor_intercept <dbl>,
#> #   sensor_range <chr>, sensor_range_min <dbl>, sensor_range_max <dbl>,
#> #   sensor_resolution <dbl>, sensor_unit <chr>, sensor_accuracy <dbl>,
#> #   sensor_transmit_ratio <int>, accelerometer_algorithm <chr>, …
get_tags(con, acoustic_tag_id = c("A69-1601-16129", "A69-1601-16130"))
#> # A tibble: 2 × 54
#>   tag_ser…¹ tag_t…² tag_s…³ senso…⁴ acous…⁵ acous…⁶ manuf…⁷ model frequ…⁸ status
#>   <chr>     <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr> <chr>   <chr> 
#> 1 1187449   acoust… animal  NA      A69-16… NA      VEMCO   V7-4x 069k    avail…
#> 2 1187450   acoust… animal  NA      A69-16… NA      VEMCO   V7-4x 069k    avail…
#> # … with 44 more variables: activation_date <dttm>,
#> #   battery_estimated_life <chr>, battery_estimated_end_date <dttm>,
#> #   length <dbl>, diameter <dbl>, weight <dbl>, floating <chr>,
#> #   archive_memory <chr>, sensor_slope <dbl>, sensor_intercept <dbl>,
#> #   sensor_range <chr>, sensor_range_min <dbl>, sensor_range_max <dbl>,
#> #   sensor_resolution <dbl>, sensor_unit <chr>, sensor_accuracy <dbl>,
#> #   sensor_transmit_ratio <int>, accelerometer_algorithm <chr>, …